input validation include in the autocorrelation function
